% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-metadata-tables.R
\name{build_generic_hmm_meta}
\alias{build_generic_hmm_meta}
\title{Build a generic system metadata table from a list of padloc models}
\usage{
build_generic_hmm_meta(hmm_protein_key)
}
\arguments{
\item{hmm_protein_key}{A \code{\link[base:data.frame]{base::data.frame()}} with two columns, \code{hmm.name}
containing the names of each physical hmm file, and \code{protein.name}
containing the corresponding protein name that each hmm is associated with.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Build a generic system metadata table from a list of padloc
models.
}
