% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-metadata-tables.R
\name{build_generic_sys_meta}
\alias{build_generic_sys_meta}
\title{Build a generic system metadata table from a list of padloc models}
\usage{
build_generic_sys_meta(model_list)
}
\arguments{
\item{model_list}{A \code{\link[base:list]{base::list()}} of padloc models.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Build a generic system metadata table from a list of padloc
models.
}
