% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-hmm-meta.R
\name{collapse_hmm_meta}
\alias{collapse_hmm_meta}
\title{Collapse HMM metadata table multi-name assignments}
\usage{
collapse_hmm_meta(hmm_meta_expanded)
}
\arguments{
\item{hmm_meta_expanded}{An expanded HMM metadata table (as read-in by
\code{\link[=read_hmm_meta]{read_hmm_meta()}}) and expanded by \code{\link[=expand_hmm_meta]{expand_hmm_meta()}}.}
}
\value{
An HMM metadata table with collasped name assignments.
}
\description{
Collapse HMM metadata table multi-name assignments.
}
\examples{
path <- padlocdev_example("padloc-db/hmm_meta.txt")
hmm_meta <- read_hmm_meta(path)
hmm_meta_expanded <- expand_hmm_meta(hmm_meta)
hmm_meta_collapsed <- collapse_hmm_meta(hmm_meta_expanded)
hmm_meta_collapsed
}
