% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-methods.R
\name{compare_hmm_files_hmm_meta}
\alias{compare_hmm_files_hmm_meta}
\title{Compare HMM files against those listed in hmm_meta.txt}
\usage{
compare_hmm_files_hmm_meta(hmm_list, hmm_meta)
}
\arguments{
\item{hmm_list}{A \code{\link[base:list]{base::list()}} of HMMs (i.e. as read in by
\code{\link[=multi_read_hmm]{multi_read_hmm()}} or \code{\link[=multi_read_hmm_header]{multi_read_hmm_header()}}).}

\item{hmm_meta}{A padloc-db HMM metadata file (i.e. as read in by
\code{\link[=read_hmm_meta]{read_hmm_meta()}}).}
}
\value{
a \code{\link[base:list]{base::list()}} of three \code{\link[tibble:tibble]{tibble::tibble()}}s.
}
\description{
Compare HMM files against those listed in hmm_meta.txt.
}
