% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-methods.R
\name{compare_sys_files_hmm_meta}
\alias{compare_sys_files_hmm_meta}
\title{Compare system model files against the hmm metadata table.}
\usage{
compare_sys_files_hmm_meta(models_expanded, hmm_meta_expanded)
}
\arguments{
\item{models_expanded}{A \code{\link[base:list]{base::list()}} of padloc-db system models (i.e. as
read in by \code{\link[=multi_read_padloc_model]{multi_read_padloc_model()}}).}

\item{hmm_meta_expanded}{A padloc-db HMM metadata file (i.e. as read in
by \code{\link[=read_hmm_meta]{read_hmm_meta()}}).}
}
\value{
A \code{\link[base:list]{base::list()}} of three \code{\link[tibble:tibble]{tibble::tibble()}}s.
}
\description{
Compare system model files against those listed in hmm_meta.txt.
}
