% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-methods.R
\name{compare_sys_files_sys_meta}
\alias{compare_sys_files_sys_meta}
\title{Compare system model files against those listed in sys_meta.txt}
\usage{
compare_sys_files_sys_meta(sys_list, sys_meta)
}
\arguments{
\item{sys_list}{A \code{\link[base:list]{base::list()}} of system models (i.e. as read in by
\code{\link[=multi_read_padloc_model]{multi_read_padloc_model()}}.}

\item{sys_meta}{A padloc-db system metadata file (i.e. as read in by
\code{\link[=read_sys_meta]{read_sys_meta()}}).}
}
\value{
a \code{\link[base:list]{base::list()}} of three \code{\link[tibble:tibble]{tibble::tibble()}}s.
}
\description{
Compare system model files against those listed in hmm_meta.txt.
}
