% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-methods.R
\name{compare_sys_meta_sys_groups}
\alias{compare_sys_meta_sys_groups}
\title{Compare system model files against those listed in sys_meta.txt}
\usage{
compare_sys_meta_sys_groups(sys_meta, sys_groups)
}
\arguments{
\item{sys_meta}{A padloc-db system metadata file (i.e. as read in by
\code{\link[=read_sys_meta]{read_sys_meta()}}).}

\item{sys_groups}{A padloc-db system groups metadata file (i.e. as read in
by \code{\link[=read_sys_groups]{read_sys_groups()}}).}
}
\value{
a \code{\link[base:list]{base::list()}} of three \code{\link[tibble:tibble]{tibble::tibble()}}s.
}
\description{
Compare system model files against those listed in hmm_meta.txt.
}
