% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{crossref_data_pull_journal}
\alias{crossref_data_pull_journal}
\title{Create a clean 'journal' column}
\usage{
crossref_data_pull_journal(crossref_data)
}
\arguments{
\item{crossref_data}{A data frame from \code{\link[=get_crossref_data]{get_crossref_data()}}.}
}
\value{
A data frame.
}
\description{
Create a clean 'journal' column
}
