% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide-database.R
\name{divide_database}
\alias{divide_database}
\title{divide_database}
\usage{
divide_database(
  sys_expanded,
  sys_groups,
  hmm_meta_expanded,
  sys_meta,
  path,
  new_path
)
}
\arguments{
\item{sys_expanded}{...}

\item{sys_groups}{...}

\item{hmm_meta_expanded}{...}

\item{sys_meta}{...}

\item{path}{...}

\item{new_path}{...}
}
\description{
divide_database.
}
