% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve-doi.R
\name{doi_to_long}
\alias{doi_to_long}
\title{Convert a short doi to the corresponding long doi}
\usage{
doi_to_long(short_doi)
}
\arguments{
\item{short_doi}{A short doi in the form '10/xxxx'.}
}
\value{
A long doi.
}
\description{
Convert a short doi to the corresponding long doi
}
\examples{
doi_to_long("10/gzgh")
}
