% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve-doi.R
\name{doi_to_long_df}
\alias{doi_to_long_df}
\title{Convert a dataframe of short dois to long dois}
\usage{
doi_to_long_df(df, col = doi_short)
}
\arguments{
\item{df}{A data frame containing a column with short dois}

\item{col}{The name of the column containing short dois}
}
\value{
A data frame with a new column containing long dois
}
\description{
Convert a dataframe of short dois to long dois
}
