% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve-doi.R
\name{doi_to_short}
\alias{doi_to_short}
\title{Convert a long doi to the corresponding short doi}
\usage{
doi_to_short(long_doi)
}
\arguments{
\item{long_doi}{A long doi in the form '10.1093/NAR/GKT1419'.}
}
\value{
A short doi.
}
\description{
Convert a long doi to the corresponding short doi
}
\examples{
doi_to_short("10.1093/NAR/GKT1419")
}
