% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-gene-groups.R
\name{expand_gene_groups}
\alias{expand_gene_groups}
\title{Expand gene groups}
\usage{
expand_gene_groups(model, gene_type, hmm_meta)
}
\arguments{
\item{model}{A padloc model.}

\item{gene_type}{The category of genes to expand, i.e. one of: "core_genes",
"optional_genes", or "prohibited_genes".}

\item{hmm_meta}{A padloc-db HMM metadata table.}
}
\value{
A \code{\link[base:character]{base::character()}} vector containing the expanded gene values.
}
\description{
Expand secondary gene assignments in a specific category of
genes into their underlying genes.
}
\examples{
# Read data -----------------------------------------------------------------
path <- padlocdev_example("padloc-db/sys/DRT_class_I.yaml")
model <- read_padloc_model(path)
model
path <- padlocdev_example("padloc-db/hmm_meta.txt") # Generate path for example model
hmm_meta <- read_hmm_meta(path)
hmm_meta
# Expand genes --------------------------------------------------------------
# The category "optional_genes" only contains the gene group "Sya_accessory"
model$optional_genes
# After expanding, "Sya_accessory" becomes "SyaB" and "SyaC"
model$optional_genes <- expand_gene_groups(model, "optional_genes", hmm_meta)
model$optional_genes
}
