% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{expand_gene_groups_all}
\alias{expand_gene_groups_all}
\title{Expand all gene groups in a list of padloc models}
\usage{
expand_gene_groups_all(models, hmm_meta)
}
\arguments{
\item{models}{A list of padloc models.}

\item{hmm_meta}{A padloc-db HMM metadata table.}
}
\value{
A \code{\link[base:list]{base::list()}} of padloc models with expanded genes.
}
\description{
Expand secondary gene assignments in all categories of genes
in a list of padloc models into their underlying genes.
}
\examples{
model <- multi_read_padloc_model(padlocdev_example("sys"))
hmm_meta <- read_hmm_meta(padlocdev_example("hmm_meta.txt"))
}
