% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide-database.R
\name{filter_hmm_meta}
\alias{filter_hmm_meta}
\title{filter_hmm_meta}
\usage{
filter_hmm_meta(sys_list, hmm_meta)
}
\arguments{
\item{sys_list}{...}

\item{hmm_meta}{...}
}
\description{
filter_hmm_meta.
}
