% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide-database.R
\name{filter_models}
\alias{filter_models}
\title{filter_models}
\usage{
filter_models(sys_list, sys_groups, group_name)
}
\arguments{
\item{sys_list}{...}

\item{sys_groups}{...}

\item{group_name}{...}
}
\description{
filter_models.
}
