% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide-database.R
\name{filter_sys_meta}
\alias{filter_sys_meta}
\title{filter_sys_meta}
\usage{
filter_sys_meta(sys_list, sys_meta)
}
\arguments{
\item{sys_list}{...}

\item{sys_meta}{...}
}
\description{
filter_sys_meta.
}
