% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{get_crossref_data}
\alias{get_crossref_data}
\title{Get a table of crossref data.}
\usage{
get_crossref_data(df, col = doi_long)
}
\arguments{
\item{df}{A data frame with a column containing long dois.}

\item{col}{The name of the column containing long dois.}
}
\value{
A data frame.
}
\description{
Get a table of crossref data.
}
