% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-wrappers.R
\name{hmm_files_vs_hmm_meta}
\alias{hmm_files_vs_hmm_meta}
\title{Compare HMM files against those listed in hmm_meta.txt}
\usage{
hmm_files_vs_hmm_meta(hmm_list, hmm_meta)
}
\arguments{
\item{hmm_list}{A \code{\link[base:list]{base::list()}} of HMMs (i.e. as read in by
\code{\link[=multi_read_hmm]{multi_read_hmm()}} or \code{\link[=multi_read_hmm_header]{multi_read_hmm_header()}}).}

\item{hmm_meta}{A padloc-db HMM metadata file (i.e. as read in by
\code{\link[=read_hmm_meta]{read_hmm_meta()}}).}
}
\value{
a \code{\link[base:list]{base::list()}} with two \href{s}{tibble::tibble()},
"HMMs without metadata" - listing HMMs that are not referenced in the
HMM metadata file, and "Metadata without HMMs" - listing HMMs that are
referenced in the HMM metadata file, but do not actually exist.
}
\description{
Compare HMM files against those listed in hmm_meta.txt.
}
