% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-hmms.R
\name{multi_read_hmm}
\alias{multi_read_hmm}
\title{Read all HMMs in a directory into a list}
\usage{
multi_read_hmm(directory)
}
\arguments{
\item{directory}{Path to a directory containing profile HMMs (*.hmm)}
}
\value{
A \code{\link[base:list]{base::list()}} of HMMs.
}
\description{
Read all HMMs in a directory into a list
}
\examples{
hmms <- multi_read_hmm(padlocdev_example("padloc-db/hmm"))
hmms[1]
}
