% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-hmms.R
\name{multi_read_hmm_header}
\alias{multi_read_hmm_header}
\title{Read the headers from a profile HMM into a list.}
\usage{
multi_read_hmm_header(directory)
}
\arguments{
\item{directory}{Path to a directory containing profile HMMs (*.hmm)}
}
\value{
A \code{\link[base:list]{base::list()}}.
}
\description{
Read the headers from a profile HMM into a list.
}
\examples{
hmms <- multi_read_hmm_header(padlocdev_example("padloc-db/hmm"))
hmms[1]
}
