% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-padloc-model.R
\name{multi_read_padloc_model}
\alias{multi_read_padloc_model}
\title{Read the filename, NAME and ACC identifiers from a profile HMM into a list.}
\usage{
multi_read_padloc_model(directory)
}
\arguments{
\item{directory}{Path to a directory containing profile HMMs (*.hmm)}
}
\value{
A \code{\link[base:list]{base::list()}}.
}
\description{
Read the filename, NAME and ACC identifiers from a profile HMM into a list.
}
\examples{
multi_read_hmm_header(padlocdev_example("padloc-db/sys"))
}
