% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-padloc-model.R
\name{multi_write_padloc_model}
\alias{multi_write_padloc_model}
\title{Write out a list of padloc model (\verb{*.yaml}) files}
\usage{
multi_write_padloc_model(model_list, directory)
}
\arguments{
\item{model_list}{A \code{\link[base:list]{base::list()}} of padloc models.}

\item{directory}{Path to a directory, where files will be written.}
}
\description{
Write out a list of padloc model (\verb{*.yaml}) files.
}
