% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-padloc-model.R
\name{padloc_model_to_chr}
\alias{padloc_model_to_chr}
\title{Convert a padloc model to a \code{\link[glue:glue]{glue::glue()}} character vector}
\usage{
padloc_model_to_chr(model)
}
\arguments{
\item{model}{A padloc model.}
}
\value{
A length 1 \code{\link[glue:glue]{glue::glue()}} character vector.
}
\description{
Convert a padloc model to a \code{\link[glue:glue]{glue::glue()}} character vector.
Used for preparing a padloc model to be written to a file.
}
