% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padlocdev-package.R
\docType{package}
\name{padlocdev-package}
\alias{padlocdev}
\alias{padlocdev-package}
\title{padlocdev: PADLOC development tools}
\description{
PADLOC development tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leightonpayne/padlocdev}
  \item \url{https://leightonpayne.github.io/padlocdev/}
  \item Report bugs at \url{https://github.com/leightonpayne/padlocdev/issues}
}

}
\author{
\strong{Maintainer}: Leighton Payne \email{leighton.payne@postgrad.otago.ac.nz} (\href{https://orcid.org/0000-0003-2305-6827}{ORCID})

}
\keyword{internal}
