% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-methods.R
\name{pull_genes}
\alias{pull_genes}
\title{Pull the genes from a list of system models}
\usage{
pull_genes(sys_list)
}
\arguments{
\item{sys_list}{A \code{\link[base:list]{base::list()}} of system models (i.e. as read in by
\code{\link[=multi_read_padloc_model]{multi_read_padloc_model()}}.}
}
\value{
a \code{\link[base:character]{base::character()}} vector.
}
\description{
Pull the genes from a list of system models.
}
