% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide-database.R
\name{pull_intersect}
\alias{pull_intersect}
\title{pull_intersect}
\usage{
pull_intersect(group_overlap, group_1, group_2)
}
\arguments{
\item{group_overlap}{...}

\item{group_1}{...}

\item{group_2}{...}
}
\description{
pull_intersect.
}
\details{
TODO: Write documentation.
}
