% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-hmms.R
\name{read_hmm}
\alias{read_hmm}
\title{Read an HMMER3 formatted profile HMM}
\usage{
read_hmm(file)
}
\arguments{
\item{file}{Path to a profile HMM generated with HMMER3}
}
\value{
A \code{\link[base:list]{base::list()}}.
}
\description{
Read an HMMER3 formatted profile HMM
}
\examples{
read_hmm(padlocdev_example("padloc-db/hmm/PDLC00150.hmm"))
}
