% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-metadata-tables.R
\name{read_hmm_meta}
\alias{read_hmm_meta}
\title{Read a padloc-db \code{hmm_meta.txt} file.}
\usage{
read_hmm_meta(file)
}
\arguments{
\item{file}{Path to sys_meta.txt}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Read a padloc-db \code{hmm_meta.txt} file.
}
\examples{
sys_meta <- read_sys_meta(padlocdev_example("padloc-db/sys_meta.txt"))
}
