% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-methods.R
\name{read_hmm_name_acc}
\alias{read_hmm_name_acc}
\title{Read the filename, NAME and ACC identifiers from a profile HMM into a list.}
\usage{
read_hmm_name_acc(file)
}
\arguments{
\item{file}{Path to profile HMM (*.hmm)}
}
\value{
A \code{\link[base:list]{base::list()}}.
}
\description{
Read the filename, NAME and ACC identifiers from a profile HMM into a list.
}
\examples{
read_hmm_name_acc(padlocdev_example("protein.hmm"))
}
