% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-methods.R
\name{read_padloc_model}
\alias{read_padloc_model}
\title{Read in a padloc model (.yaml) file}
\usage{
read_padloc_model(file)
}
\arguments{
\item{file}{Path to padloc model in yaml format (*.yaml).}
}
\value{
A \code{\link[base:list]{base::list()}}.
}
\description{
Read in a padloc model (.yaml) file
}
\examples{
padloc_model <- read_padloc_model(padlocdev_example("example_model.yaml"))
}
