% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-metadata-tables.R
\name{read_sys_groups}
\alias{read_sys_groups}
\title{Read a padloc-db sys_groups.txt file.}
\usage{
read_sys_groups(file)
}
\arguments{
\item{file}{Path to sys_groups.txt}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Read a padloc-db sys_groups.txt file.
}
\examples{
sys_groups <- read_sys_groups(padlocdev_example("padloc-db/sys_groups.txt"))
sys_groups
}
