% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-metadata-tables.R
\name{read_sys_master}
\alias{read_sys_master}
\title{Read a padloc-db sys_master.txt file.}
\usage{
read_sys_master(file)
}
\arguments{
\item{file}{Path to sys_master.txt.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Read a padloc-db sys_master.txt file.
}
