% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padloc-model.R
\name{report_padloc_model_validity}
\alias{report_padloc_model_validity}
\title{Generate a validity report for a list of padloc models}
\usage{
report_padloc_model_validity(models)
}
\arguments{
\item{models}{A \code{\link[base:list]{base::list()}} of system models (i.e. as read in by
\code{\link[=multi_read_padloc_model]{multi_read_padloc_model()}}.}
}
\value{
A \code{\link[purrr:safely]{purrr::safely()}} style \code{\link[base:list]{base::list()}} with components \code{result}
and \code{error}. If the model is valid, \code{result} is \code{TRUE} and \code{error} is \code{NULL}.
If the model is invalid, \code{result} is \code{FALSE} and \code{error} is an \code{error} object
that describes the reason the model is invalid.
}
\description{
Generate a validity report for a list of padloc models.
}
