% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide-database.R
\name{summarise_groups}
\alias{summarise_groups}
\title{summarise_groups}
\usage{
summarise_groups(sys_expanded, sys_groups)
}
\arguments{
\item{sys_expanded}{...}

\item{sys_groups}{...}
}
\description{
summarise_groups
}
