% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-system.R
\name{system_cp_files}
\alias{system_cp_files}
\title{Copy a list of files on the system}
\usage{
system_cp_files(file_list, extension, path, new_path)
}
\arguments{
\item{file_list}{A character vector of file names (without extensions).}

\item{extension}{The extension to apply to file names.}

\item{path}{The path to the origin directory containing the files to copy.}

\item{new_path}{The path to the destination directory. If it doesn't exist,
this directory will be created (recursively).}
}
\value{
Nothing.
}
\description{
Copy a list of files on the system.
}
