% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-system.R
\name{system_cp_hmm}
\alias{system_cp_hmm}
\title{Copy profile HMM files from one directory to another}
\usage{
system_cp_hmm(hmm_list, path, new_path)
}
\arguments{
\item{hmm_list}{A character vector listing HMM file names to be copied, such
as that generated by...}

\item{path}{The path to}

\item{new_path}{asd...}
}
\value{
Nothing.
}
\description{
Copy profile HMM files from one directory to another.
}
