% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-system.R
\name{system_cp_sys}
\alias{system_cp_sys}
\title{Copy system description files from one directory to another}
\usage{
system_cp_sys(sys_list, path, new_path)
}
\arguments{
\item{sys_list}{A character vector listing system description file names to
be copied, such as that generated by...}

\item{path}{The path to}

\item{new_path}{asd...}
}
\value{
Nothing.
}
\description{
Copy system description files from one directory to another.
}
