% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-system.R
\name{system_mv_deprecated}
\alias{system_mv_deprecated}
\title{Move system description files to deprecated directory}
\usage{
system_mv_deprecated(path)
}
\arguments{
\item{path}{The path to padloc-db/sys/}
}
\value{
Nothing.
}
\description{
Move system description files to deprecated directory
}
