% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padloc-model.R
\name{valid_padloc_model_basic}
\alias{valid_padloc_model_basic}
\title{Basic validity checker for a padloc model}
\usage{
valid_padloc_model_basic(padloc_model)
}
\arguments{
\item{padloc_model}{A padloc model.}
}
\value{
\code{TRUE} if the model is valid, else an \code{error} object.
}
\description{
Basic validity checker for a padloc model.
}
