% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padloc-model.R
\name{valid_padloc_model_expanded}
\alias{valid_padloc_model_expanded}
\title{Validity checker for a padloc model with expanded genes}
\usage{
valid_padloc_model_expanded(padloc_model)
}
\arguments{
\item{padloc_model}{A padloc model.}
}
\value{
\code{TRUE} if the model is valid, else an \code{error} object.
}
\description{
Validity checker for a padloc model with expanded genes.
}
