% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-methods.R
\name{which_uses}
\alias{which_uses}
\title{Find out which system models use particular proteins}
\usage{
which_uses(genes, models_expanded)
}
\arguments{
\item{genes}{Character aray of gene names.}

\item{models_expanded}{List of expanded models.}
}
\value{
A list of system models.
}
\description{
Find out which system models use particular proteins.
}
