% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padloc-model.R
\name{why_invalid}
\alias{why_invalid}
\title{Determine why padloc models are invalid}
\usage{
why_invalid(validity_report)
}
\arguments{
\item{validity_report}{A validity report as generated by
\code{\link[=report_padloc_model_validity]{report_padloc_model_validity()}}.}
}
\value{
A named list of invalid models and the associated error.
}
\description{
Determine why padloc models are invalid.
}
