% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-padloc-model.R
\name{write_padloc_model}
\alias{write_padloc_model}
\title{Write out a padloc model (\verb{*.yaml}) file}
\usage{
write_padloc_model(model, file)
}
\arguments{
\item{model}{A padloc model.}

\item{file}{Path to a file.}
}
\description{
Write out a padloc model (\verb{*.yaml}) file.
}
