% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padloc-model.R
\name{padloc_model}
\alias{padloc_model}
\title{Padloc model}
\usage{
padloc_model(
  maximum_separation,
  minimum_core,
  minimum_total,
  core_genes,
  optional_genes,
  prohibited_genes
)
}
\arguments{
\item{maximum_separation}{A \code{\link[base:integer]{base::integer()}}, representing the maximum
number of unrelated genes that can seperate the genes of a system.}

\item{minimum_core}{A \code{\link[base:integer]{base::integer()}}, representing the minimum number
of core genes that must be identified for a complete system.}

\item{minimum_total}{A \code{\link[base:integer]{base::integer()}}, representing the minimum number
of total genes that must be identified for a complete system.}

\item{core_genes}{A \code{\link[base:character]{base::character()}}, listing the names of the core genes
that contribute to system completeness.}

\item{optional_genes}{A \code{\link[base:character]{base::character()}}, listing the names of the
optional genes that contribute to system completeness.}

\item{prohibited_genes}{A \code{\link[base:character]{base::character()}}, listing the names of the
prohibited genes that cannot be identified in proximity to the system.}
}
\value{
A \code{\link[base:list]{base::list()}}.
}
\description{
A list of all the parameters that make up a padloc model.
}
\examples{
model <- padloc_model(
  maximum_separation = 3,
  minimum_core = 2,
  minimum_total = 3,
  core_genes = c("GenA", "GenB"),
  optional_genes = c("GenC", "GenD"),
  prohibited_genes = "NA"
)
}
